feat: Enhance Filament Copilot Plugin with new features and tools
- Added support for chat history management with `chatHistoryEnabled` and `isChatHistoryEnabled` methods.
- Introduced streaming capabilities with `streaming`, `isStreamingEnabled`, and `streamingChunkSize` methods.
- Implemented export functionality with `exportEnabled`, `isExportEnabled`, `exportFormats`, and `getExportFormats` methods.
- Added token budget management with `tokenBudgetEnabled`, `isTokenBudgetEnabled`, `dailyTokenBudget`, and `monthlyTokenBudget` methods.
- Introduced rate limiting features with `rateLimitEnabled` and `isRateLimitEnabled` methods.
- Added memory management with `memoryEnabled`, `isMemoryEnabled`, and `maxMemoriesPerUser` methods.
- Implemented authorization respect feature with `respectAuthorization` and `shouldRespectAuthorization` methods.
- Enhanced `StreamController` to handle new streaming events and tool calls.
- Updated `CopilotChat` to manage thinking messages and tool call results.
- Added `ExecuteBulkActionTool` for executing bulk actions on multiple records.
- Improved existing tools to support eager loading of relationships and counts.
- Updated tests to cover new features and ensure proper functionality.

+++ src/Widgets/CopilotStatsOverview.php
@@ -25,14 +25,14 @@ class CopilotStatsOverview extends StatsOverviewWidget
             Stat::make(
                 __('filament-copilot::filament-copilot.tokens_today'),
                 number_format(
-                    CopilotTokenUsage::where('usage_date', $today)->sum('total_tokens')
+                    (int) CopilotTokenUsage::where('usage_date', $today)->sum('total_tokens')
                 ),
             )->icon('heroicon-o-calculator'),
 
             Stat::make(
                 __('filament-copilot::filament-copilot.tokens_this_month'),
                 number_format(
-                    CopilotTokenUsage::where('usage_date', '>=', now()->startOfMonth()->toDateString())->sum('total_tokens')
+                    (int) CopilotTokenUsage::where('usage_date', '>=', now()->startOfMonth()->toDateString())->sum('total_tokens')
                 ),
             )->icon('heroicon-o-chart-bar'),
